+++ sdk/core/azure-core/inc/azure/core/http/http.hpp
@@ -525,14 +525,17 @@ namespace Azure { namespace Core { namespace Http {
     explicit Request(HttpMethod httpMethod, Url url);
 
     /**
-     * @brief Add HTTP header to the #Azure::Core::Http::Request.
+     * @brief Set an HTTP header to the #Azure::Core::Http::Request.
      *
-     * @param name The name for the header to be added.
-     * @param value The value for the header to be added.
+     * @remark If the header key does not exists, it is added.
+     *
+     *
+     * @param name The name for the header to be set or added.
+     * @param value The value for the header to be set or added.
      *
      * @throw if \p name is an invalid header key.
      */
-    void AddHeader(std::string const& name, std::string const& value);
+    void SetHeader(std::string const& name, std::string const& value);
 
     /**
      * @brief Remove an HTTP header.
@@ -660,19 +663,19 @@ namespace Azure { namespace Core { namespace Http {
     // ===== Methods used to build HTTP response =====
 
     /**
-     * @brief Add HTTP header to the #Azure::Core::Http::RawResponse.
+     * @brief Set an HTTP header to the #Azure::Core::Http::Request.
      *
      * @remark The \p name must contain valid header name characters (RFC 7230).
      *
-     * @param name The name for the header to be added.
-     * @param value The value for the header to be added.
+     * @param name The name for the header to be set or added.
+     * @param value The value for the header to be set or added.
      *
      * @throw if \p name contains invalid characters.
      */
-    void AddHeader(std::string const& name, std::string const& value);
+    void SetHeader(std::string const& name, std::string const& value);
 
     /**
-     * @brief Add HTTP header to the #Azure::Core::Http::RawResponse.
+     * @brief Set an HTTP header to the #Azure::Core::Http::Request.
      *
      * @remark The \p header must contain valid header name characters (RFC 7230).
      * @remark Header name, value and delimiter are expected to be in \p header.
@@ -681,10 +684,10 @@ namespace Azure { namespace Core { namespace Http {
      *
      * @throw if \p header has an invalid header name or if the delimiter is missing.
      */
-    void AddHeader(std::string const& header);
+    void SetHeader(std::string const& header);
 
     /**
-     * @brief Add HTTP header to the #Azure::Core::Http::RawResponse.
+     * @brief Set an HTTP header to the #Azure::Core::Http::Request.
      *
      * @remark The string referenced by \p first and \p last must contain valid header name
      * characters (RFC 7230).
@@ -697,7 +700,7 @@ namespace Azure { namespace Core { namespace Http {
      * @throw if the string referenced by \p first and \p last contains an invalid header name or if
      * the delimiter is missing.
      */
-    void AddHeader(uint8_t const* const first, uint8_t const* const last);
+    void SetHeader(uint8_t const* const first, uint8_t const* const last);
 
     /**
      * @brief Set #Azure::IO::BodyStream for this HTTP response.