+++ sdk/core/azure-core/src/http/raw_response.cpp
@@ -18,7 +18,7 @@ std::string const& RawResponse::GetReasonPhrase() const { return m_reasonPhrase;
 
 Azure::Core::CaseInsensitiveMap const& RawResponse::GetHeaders() const { return this->m_headers; }
 
-void RawResponse::AddHeader(uint8_t const* const first, uint8_t const* const last)
+void RawResponse::SetHeader(uint8_t const* const first, uint8_t const* const last)
 {
   // get name and value from header
   auto start = first;
@@ -40,17 +40,17 @@ void RawResponse::AddHeader(uint8_t const* const first, uint8_t const* const las
   end = std::find(start, last, '\r');
   auto headerValue = std::string(start, end); // remove \r
 
-  AddHeader(headerName, headerValue);
+  SetHeader(headerName, headerValue);
 }
 
-void RawResponse::AddHeader(std::string const& header)
+void RawResponse::SetHeader(std::string const& header)
 {
-  return AddHeader(
+  return SetHeader(
       reinterpret_cast<uint8_t const*>(header.data()),
       reinterpret_cast<uint8_t const*>(header.data() + header.size()));
 }
 
-void RawResponse::AddHeader(std::string const& name, std::string const& value)
+void RawResponse::SetHeader(std::string const& name, std::string const& value)
 {
   return Details::InsertHeaderWithValidation(this->m_headers, name, value);
 }